moved the email route to its own router and file
