+++ server/routes/order.router.js
@@ -6,19 +6,6 @@ const {
   rejectUnauthenticated,
 } = require('../modules/authentication-middleware');
 
-// import nodemailer
-const nodemailer = require("nodemailer");
-// make the "transporter"
-// this is the email that will do the sending 
-const transporter = nodemailer.createTransport({
-  service: "outlook",
-  auth: {
-    user: process.env.EMAIL,
-    pass: process.env.PASSWORD
-  },
-});
-
-
 
 router.get('/', rejectUnauthenticated, async (req, res) => {
   try {
@@ -102,31 +89,6 @@ router.post('/initialSample', rejectUnauthenticated, async (req, res) => {
   }
 });
 
-
-/*
-the email needs a req.body as follows:
-{
-  customerEmail: users.email
-  message: 'The message to be sent'
-}
-*/
-router.post('/email', rejectUnauthenticated, async (req, res) => {
-  const info =  await transporter.sendMail({
-    from: process.env.EMAIL,
-    to: `${req.body.customerEmail}`,
-    subject: "Sample info from Canomiks",
-    text: `${req.body.message}`,
-  }, (err, info) => {
-    if (err) {
-      console.log('💥 error sending', err);
-      return;
-    } ;
-    console.log('🎉 it has been sent', info.response)
-  });
-  
-  res.sendStatus(200);
-})
-
 // for add sample page to save the sample information; after initial insert
 router.put('/updateOrder', rejectUnauthenticated, async (req, res) => {
   try {
@@ -159,7 +121,6 @@ router.put('/updateOrder', rejectUnauthenticated, async (req, res) => {
 
 // for shipping page to save the shipping information; after initial insert
 router.put('/shipping', rejectUnauthenticated, async (req, res) => {
-  console.log('🎺 made it to the put:', req.body)
   // POST route code here
   try {
     const order = req.body;
