+++ src/redux/sagas/shipping.saga.js
@@ -2,14 +2,13 @@ import { put, takeLatest } from 'redux-saga/effects';
 import axios from 'axios';
 
 function* addShippingInfo(action) {
-  console.log('🎸 in edit shipping', action.payload);
   try {
     const response = yield axios.put('/api/orders/shipping', action.payload);
-    console.log('🎻 response is:', response.data);
 
-    // yield put({
-    //   type: 'EDIT_SHIPPING',
-    // });
+    yield put({
+      type: 'EDIT_SHIPPING',
+      payload: response.data
+    });
 
   } catch (err) {
     console.log('💥 error in shipping', err);
